requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cloudinary==1.28.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==35.0.0
defusedxml==0.7.1
dj-database-url==0.5.0
dj-rest-auth==2.1.11
Django==3.2.9
django-allauth==0.44.0
django-filter==21.1
django-on-heroku==1.1.2
django-rest-swagger==2.2.0
django-utils-six==2.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
httpie==2.6.0
idna==3.3
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
oauthlib==3.1.1
openapi-codec==1.3.2
psycopg2-binary==2.9.1
pycparser==2.21
Pygments==2.10.0
PyJWT==2.3.0
PySocks==1.7.1
python-dotenv==0.19.2
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
simplejson==3.17.5
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.7
whitenoise==5.3.0
